+++ consoleapp/FibonacciConsoleApp/FibonacciConsoleAppTests/FibonacciMultiplicationTableTests.cs
@@ -12,10 +12,10 @@ namespace FibonacciMultiplicationTableConsoleAppTests
         [TestMethod]
         public void FibonacciMultiplicationTableZeroElements()
         {
-            BigInteger first = 0;
-            BigInteger second = 1;
+            Int64 first = 0;
+            Int64 second = 1;
             int sequenceLength = 0;
-            List<List<BigInteger>> fibonacciMultiplicationTableActual = new List<List<BigInteger>>();
+            List<List<Int64>> fibonacciMultiplicationTableActual = new List<List<Int64>>();
             fibonacciMultiplicationTableActual = Fibonacci.GetMultiplicationTable(first, second, sequenceLength);
 
             Assert.AreEqual(0, fibonacciMultiplicationTableActual.Count, 0, "Count of Fibonacci multiplication table with 0 elements is more than 0. Actual count " + fibonacciMultiplicationTableActual.Count);
@@ -25,7 +25,7 @@ namespace FibonacciMultiplicationTableConsoleAppTests
         public void FibonacciMultiplicationTableZeroElementsDefaults()
         {
             int sequenceLength = 0;
-            List<List<BigInteger>> fibonacciMultiplicationTableActual = new List<List<BigInteger>>();
+            List<List<Int64>> fibonacciMultiplicationTableActual = new List<List<Int64>>();
             fibonacciMultiplicationTableActual = Fibonacci.GetMultiplicationTable(sequenceLength);
 
             Assert.AreEqual(0, fibonacciMultiplicationTableActual.Count, 0, "Count of Fibonacci multiplication table defaults with 0 elements is more than 0. Actual count " + fibonacciMultiplicationTableActual.Count);
@@ -34,24 +34,24 @@ namespace FibonacciMultiplicationTableConsoleAppTests
         [TestMethod]
         public void FibonacciMultiplicationTableFiveElements()
         {
-            BigInteger first = 1;
-            BigInteger second = 1;
+            Int64 first = 1;
+            Int64 second = 1;
             int sequenceLength = 5;
-            List<List<BigInteger>> fibonacciMultiplicationTableActual = new List<List<BigInteger>>();
+            List<List<Int64>> fibonacciMultiplicationTableActual = new List<List<Int64>>();
             fibonacciMultiplicationTableActual = Fibonacci.GetMultiplicationTable(first, second, sequenceLength);
 
-            List<List<BigInteger>> fibonacciMultiplicationTableExpected = new List<List<BigInteger>>()
-            { new List<BigInteger>(){1,1,2,3,5},
-            new List<BigInteger>(){1,1,1,2,3,5},
-            new List<BigInteger>(){1,1,1,2,3,5},
-            new List<BigInteger>(){2,2,2,4,6,10},
-            new List<BigInteger>(){3,3,3,6,9,15},
-            new List<BigInteger>(){5,5,5,10,15,25}};
+            List<List<Int64>> fibonacciMultiplicationTableExpected = new List<List<Int64>>()
+            { new List<Int64>(){1,1,2,3,5},
+            new List<Int64>(){1,1,1,2,3,5},
+            new List<Int64>(){1,1,1,2,3,5},
+            new List<Int64>(){2,2,2,4,6,10},
+            new List<Int64>(){3,3,3,6,9,15},
+            new List<Int64>(){5,5,5,10,15,25}};
 
             for (int i = 0; i < fibonacciMultiplicationTableExpected.Count; i++)
             {
-                List<BigInteger> currentExpected = fibonacciMultiplicationTableExpected[i];
-                List<BigInteger> currentActual = fibonacciMultiplicationTableActual[i];
+                List<Int64> currentExpected = fibonacciMultiplicationTableExpected[i];
+                List<Int64> currentActual = fibonacciMultiplicationTableActual[i];
 
                 CollectionAssert.AreEqual(currentExpected, currentActual,
                 "Expected result of Fibonacci multiplication table with 5 elements is not correct at row " + i + ". " +
@@ -64,21 +64,21 @@ namespace FibonacciMultiplicationTableConsoleAppTests
         public void FibonacciMultiplicationTableFiveElementsDefaults()
         {
             int sequenceLength = 5;
-            List<List<BigInteger>> fibonacciMultiplicationTableActual = new List<List<BigInteger>>();
+            List<List<Int64>> fibonacciMultiplicationTableActual = new List<List<Int64>>();
             fibonacciMultiplicationTableActual = Fibonacci.GetMultiplicationTable(sequenceLength);
 
-            List<List<BigInteger>> fibonacciMultiplicationTableExpected = new List<List<BigInteger>>()
-            { new List<BigInteger>(){0,1,1,2,3},
-            new List<BigInteger>(){0,0,0,0,0,0},
-            new List<BigInteger>(){1,0,1,1,2,3},
-            new List<BigInteger>(){1,0,1,1,2,3},
-            new List<BigInteger>(){2,0,2,2,4,6},
-            new List<BigInteger>(){3,0,3,3,6,9} };
+            List<List<Int64>> fibonacciMultiplicationTableExpected = new List<List<Int64>>()
+            { new List<Int64>(){0,1,1,2,3},
+            new List<Int64>(){0,0,0,0,0,0},
+            new List<Int64>(){1,0,1,1,2,3},
+            new List<Int64>(){1,0,1,1,2,3},
+            new List<Int64>(){2,0,2,2,4,6},
+            new List<Int64>(){3,0,3,3,6,9} };
 
             for (int i = 0; i < fibonacciMultiplicationTableExpected.Count; i++)
             {
-                List<BigInteger> currentExpected = fibonacciMultiplicationTableExpected[i];
-                List<BigInteger> currentActual = fibonacciMultiplicationTableActual[i];
+                List<Int64> currentExpected = fibonacciMultiplicationTableExpected[i];
+                List<Int64> currentActual = fibonacciMultiplicationTableActual[i];
 
                 CollectionAssert.AreEqual(currentExpected, currentActual,
                 "Expected result of Fibonacci multiplication table with 5 elements defaults is not correct at row " + i + ". " +
